Volume control.

+++ src/config.h
@@ -33,4 +33,4 @@
 #define STACK_SIZE_RFID 0x0800
 #define STACK_SIZE_GPIO 0x0800
 
-#define DEBOUNCE_DELAY (static_cast<uint32_t>(50))
+#define DEBOUNCE_DELAY 50
